requirements.txt
fastapi
uvicorn
aws-cdk-lib
constructs
mangum
